fix bugs

+++ frontend/src/hooks/useImportDialog.js
@@ -22,34 +22,37 @@ const useImportDialog = (refreshJobApps, handleOpenNotification) => {
     setLoading(true);
     const data = new FormData(event.currentTarget);
     const oldUserId = getUserIdFromLocalStorage();
-    const newUserId = data.get("userId");
+    const newUserId = data.get("userId").trim();
 
-    await getJobApps(newUserId)
-      .then(() => {
-        storeUserIdFromLocalStorage(newUserId);
+    if (oldUserId === newUserId) {
+      handleOpenNotification("You are already on your account!", 3000, "error");
+      setLoading(false);
+    } else {
+      await getJobApps(newUserId)
+        .then((res) => {
+          storeUserIdFromLocalStorage(newUserId);
+          deleteUser(oldUserId);
 
-        deleteUser(oldUserId);
-
-        refreshJobApps().then(() => {
+          refreshJobApps().then(() => {
+            handleOpenNotification(
+              "Successfully imported data!",
+              2000,
+              "success"
+            );
+          });
+        })
+        .catch((err) => {
+          // setError(true);
           handleOpenNotification(
-            "Successfully imported data!",
-            2000,
-            "success"
+            "Error importing data. Please check if the user ID is correct, or if your connection is stable!",
+            5000,
+            "error"
           );
+          console.error(err);
         });
-      })
-      .catch((err) => {
-        // setError(true);
-        handleOpenNotification(
-          "Error importing data. Please check if the user ID is correct, or if your connection is stable!",
-          5000,
-          "error"
-        );
-        console.error(err);
-      });
-
-    setLoading(false);
-    handleClose();
+      setLoading(false);
+      handleClose();
+    }
   };
 
   return {
